issue #1: restructure + docker-compose
