+++ Dockerfile
@@ -4,8 +4,8 @@ FROM python:3.8-slim-buster
 # Set the working directory in the container to /app
 WORKDIR /app
 
-# Copy the current directory contents into the container at /app
-ADD . /app
+# Copy the requirements.txt file into the Docker image
+COPY requirements.txt .
 
 # Install any needed packages specified in requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
@@ -18,5 +18,4 @@ RUN apt-get update && \
 # Make port 80 available to the world outside this container
 EXPOSE 80
 
-# Run bot.py when the container launches
-CMD ["python", "bot.py"]
+CMD python main.py
