+++ docker-compose.yml
@@ -1,0 +1,6 @@
+version: '3'
+services:
+  your-service:
+    build: .
+    volumes:
+      - .:/app
